requirements.txt
fastapi
uvicorn
pydantic
llama-index
llama-index-llms-openai
llama-index-retrievers-bm25
python-dotenv
numpy
pypdf
pytest
httpx
llama-index-llms-openrouter
llama-index-embeddings-huggingface
